requirements.txt
transformers
datasets
huggingface_hub
sentence-transformers
accelerate
evaluate

torch
torchvision
torchaudio

faiss-cpu
scikit-learn
numpy
pandas
matplotlib
seaborn

jupyterlab
ipywidgets

langchain
langchain-community

gradio
streamlit

opencv-python
Pillow

soundfile
librosa

python-dotenv
tqdm
requests